zircon/frontend/requirements.txt
Django==1.7
django-bootstrap3==4.11.0
django-bower==5.0.1
django-extensions==1.3.11
gevent-socketio==0.3.6
